requirements.txt
fastapi
python-telegram-bot
uvicorn
python-dotenv